+++ src/common/utils/index.js
@@ -5,5 +5,6 @@ export * from './getReceivingPieceItemStatus';
 export * from './getHydratedPieces';
 export * from './handleCommonErrors';
 export * from './handleReceiveErrorResponse';
+export * from './isSyntheticUser';
 export * from './receive';
 export * from './utils';