+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Display "Interval" column in the piece status logs. Refs UIREC-312.
 * Improve UI error when receipt fails if staff are not members of acquisitions unit. Refs UIREC-275.
 * Fix linter issues across the module. Refs UIREC-309.
+* Handle synthetic user in the piece status change log. Refs UIREC-330.
 
 ## [4.0.0](https://github.com/folio-org/ui-receiving/tree/v4.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v3.0.0...v4.0.0)